screenshots support
git-svn-id: http://crashrpt.googlecode.com/svn@493 821a4644-a21a-0410-aec7-6199d47568cc

+++ branches/1.2/src/crprober/src/main.cpp
@@ -334,7 +334,7 @@ int process_report(LPTSTR szInput, LPTSTR szInputMD5, LPTSTR szOutput,
       size_t pos = sMD5DirName.rfind('\\');
       if(pos<0) // There is no back slash in path
         sMD5DirName = _T(""); 
-      else if(pos!=(int)sMD5DirName.length()-1) // Append the back slash to dir name
+      else if(pos!=sMD5DirName.length()-1) // Append the back slash to dir name
         sMD5DirName = sMD5DirName.substr(0, pos+1);
     }
   }
